perf: 优化查询执行器和任务管理器的性能与可靠性
- **QueryExecutor**: 添加 ExecutionPlan LRU 缓存，避免重复构建查询计划
- **RegexEngine**: 实现智能引擎选择算法，根据模式复杂度自动选择最佳引擎
  - 添加 regex_complexity_score() 函数评估正则复杂度
  - 改进 lookaround 检测逻辑
  - 完善引擎选择策略的测试用例
- **TaskManager**: 将 TaskInfo.version 从 u32 升级为 u64，防止长时间运行时溢出

测试结果: 560 个测试全部通过

+++ log-analyzer/src-tauri/src/task_manager/mod.rs
@@ -67,7 +67,9 @@ pub struct TaskInfo {
     pub progress: u8,
     pub message: String,
     pub status: TaskStatus,
-    pub version: u32,
+    /// 版本号 - 使用 u64 防止长时间运行时溢出
+    /// 单调递增，用于前端幂等性检查
+    pub version: u64,
     pub workspace_id: Option<String>,
     #[serde(skip)]
     pub created_at: Instant,
@@ -194,7 +196,7 @@ impl TaskManagerActor {
                     progress: 0,
                     message: "Starting...".to_string(),
                     status: TaskStatus::Running,
-                    version: 1,
+                    version: 1u64, // 使用 u64 字面量
                     workspace_id,
                     created_at: Instant::now(),
                     completed_at: None,
